Refactored solvers Now following OCP Added info logging to all solvers
